Update and rename README.md to Elimuhubtuitioncentre.md
# Welcome to Elimuhub Tuition Centre Blog

## Who We Are

Elimuhub Tuition Centre is an educational consultancy firm located in South C, Nairobi, Kenya. We offer a range of services including home-based tuition, homeschooling, and student mentorship. We have over 100 qualified and experienced tutors ready to be deployed to clients. We also link parents to Muslim schools, IGCSE schools, and special schools. We provide placement services to teachers in our available vacancies. We link students to renowned foreign universities abroad across the world including Australia, USA, Canada, UK, and other overseas countries.

## What We Do

We offer tuition in all subjects at hourly, weekly, or monthly rates. We provide a free educational assessment to identify the needs of a student. We are passionate about education and we believe that every student deserves the best quality of education possible. We are committed to providing personalized and customized solutions to meet the needs of each student. We are also dedicated to ensuring that our tutors are well-trained and equipped with the necessary skills and knowledge to deliver effective and engaging lessons.

## How To Contact Us

We invite you to explore our website and learn more about our services and programs. You can also contact us anytime if you have any questions or inquiries. We would love to hear from you and assist you in any way we can.

You can reach us through the following channels:

- Phone: +254 731838387
- Email: info@elimuhubconsultant.co.ke 
- Facebook: (Elimuhub Tuition Centre) [https://www.facebook.com/elimuhubtuitionandconsultants] 
- Twitter: @ElimuhubK

## How Much We Charge

We offer affordable and competitive rates for our services. Our rates vary depending on the subject, level, duration, and location of the tuition. Here are some of our rates for your reference:

- Primary school tuition: Ksh 500 per hour
- Secondary school tuition: Ksh 800 per hour
- IGCSE tuition: Ksh 1000 per hour
- University tuition: Ksh 1500 per hour
- Homeschooling: Ksh 2000 per week
- Student mentorship: Ksh 3000 per month

We also offer discounts for referrals, long-term contracts, and group sessions.

Thank you for choosing Elimuhub Tuition Centre as your partner in education. We look forward to serving you and helping you achieve your academic goals.